Move data import into extra module
